+++ src/main/java/de/srendi/advancedperipherals/AdvancedPeripherals.java
@@ -27,6 +27,7 @@ import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 import org.apache.logging.log4j.Level;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.jetbrains.annotations.NotNull;
 import top.theillusivec4.curios.api.SlotTypeMessage;
 
 @Mod(AdvancedPeripherals.MOD_ID)
@@ -37,6 +38,7 @@ public class AdvancedPeripherals {
     public static final ItemGroup TAB = new ItemGroup("advancedperipheralstab") {
 
         @Override
+        @NotNull
         public ItemStack makeIcon() {
             return new ItemStack(Blocks.CHAT_BOX.get());
         }