+++ src/main/java/de/srendi/advancedperipherals/common/data/RecipesProvider.java
@@ -12,6 +12,7 @@ import net.minecraft.potion.Potions;
 import net.minecraft.tags.ItemTags;
 import net.minecraftforge.common.Tags;
 import net.minecraftforge.common.crafting.conditions.IConditionBuilder;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.function.Consumer;
 
@@ -24,7 +25,7 @@ public class RecipesProvider extends RecipeProvider implements IConditionBuilder
     }
 
     @Override
-    protected void buildShapelessRecipes(Consumer<IFinishedRecipe> consumer) {
+    protected void buildShapelessRecipes(@NotNull Consumer<IFinishedRecipe> consumer) {
         ShapedRecipeBuilder.shaped(Blocks.AR_CONTROLLER.get())
                 .define('E', Tags.Items.ENDER_PEARLS)
                 .define('C', CASING)