+++ CHANGELOG.md
@@ -4,7 +4,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.7.3r]
+
+### Removed
+- Removed `listCraftableItems()` from the RSBridge because of some issues
+
+### Added
+
+- `getMaxItemDiskStorage()` to the RS Bridge
+- `getMaxFluidDiskStorage()` to the RS Bridge
+- `getMaxItemExternalStorage()` to the RS Bridge
+- `getMaxFluidExternalStorage()` to the RS Bridge
+- `getPattern()` to the RS Bridge
+- `isItemCraftable()` to the RS Bridge
+- `isItemCraftable()` to the ME Bridge
+- french translation
+
+### Fixed
+
+- [#177]Fixed that items after removing it from a chest using the inventory manager does not stack anymore.
+- 
 
 ## [0.7.2r] - 2021-09-06